Atualização para ckeckout

+++ dataStructures/matrix/matrixMngr.c
@@ -1,5 +1,5 @@
 /*
-*	Parte do programa responsável por gerenciar a criação, manipulação e exclusão
+*	Parte do programa responsï¿½vel por gerenciar a criaï¿½ï¿½o, manipulaï¿½ï¿½o e exclusï¿½o
 *	de matrizes
 */
 #include <stdio.h>
@@ -8,7 +8,7 @@
 #include "../../datadefine.h"
 
 
-//Função para criar uma matriz a partir dos valores das m linhas e n colunas
+//Funï¿½ï¿½o para criar uma matriz a partir dos valores das m linhas e n colunas
 bool createMatrix(MatrixDescriber *matrix, unsigned int m, unsigned int n){
 
     matrix->diagNum = (m + n) -1;
@@ -41,7 +41,7 @@ bool createMatrix(MatrixDescriber *matrix, unsigned int m, unsigned int n){
     return true;
 }
 
-//Função para verificar se uma coordenada existe dentro de uma matriz
+//Funï¿½ï¿½o para verificar se uma coordenada existe dentro de uma matriz
 bool coordIsInsideMatrix(MatrixDescriber mx, Coords pos){
     if(pos.mpos >= mx.m || pos.npos >= mx.n){
         return false;
@@ -49,7 +49,7 @@ bool coordIsInsideMatrix(MatrixDescriber mx, Coords pos){
     return true;
 }
 
-//Função que dada uma matriz e uma coordenada retorna a coordenada do próximo elemento no sentido da diagonal principal
+//Funï¿½ï¿½o que dada uma matriz e uma coordenada retorna a coordenada do prï¿½ximo elemento no sentido da diagonal principal
 bool getNextElemPos(MatrixDescriber mxd, Coords * excCoord){
     excCoord->npos++;
     excCoord->mpos++;
@@ -63,7 +63,7 @@ bool getNextElemPos(MatrixDescriber mxd, Coords * excCoord){
     return true;
 }
 
-//Função que retorna um elemento da matriz dadas suas coordenadas
+//Funï¿½ï¿½o que retorna um elemento da matriz dadas suas coordenadas
 void getElement(MatrixDescriber mxd, Coords coord, float *rsp){
     /*if(!coordIsInsideMatrix(mxd, coord)){
         return false;
@@ -73,8 +73,8 @@ void getElement(MatrixDescriber mxd, Coords coord, float *rsp){
     *(rsp) = *(*(mxd.data + coord.mpos)+coord.npos);
 }
 
-//Função que converte o número de uma diagonal para suas coordenadas
-bool diagNumToCoord(MatrixDescriber mxd, unsigned int coordNum, Coords *rsp){
+//Funï¿½ï¿½o que converte o nï¿½mero de uma diagonal para suas coordenadas
+bool diagNumToCoord(MatrixDescriber mxd, unsigned short int coordNum, Coords *rsp){
     if(coordNum > mxd.diagNum-1 || coordNum < 0){
         return false;
     }
@@ -92,7 +92,7 @@ bool diagNumToCoord(MatrixDescriber mxd, unsigned int coordNum, Coords *rsp){
 }
 
 
-//Função que preenche toda a matriz com um valor inserido
+//Funï¿½ï¿½o que preenche toda a matriz com um valor inserido
 void fillMatrix(MatrixDescriber mx, float fillValue){
     register int i, j;
     for(i = 0; i < mx.m; i++){
@@ -101,7 +101,7 @@ void fillMatrix(MatrixDescriber mx, float fillValue){
         }
     }
 }
-//Função que preenche toda a matriz com valores aleatórios
+//Funï¿½ï¿½o que preenche toda a matriz com valores aleatï¿½rios
 void fillMatrixWithRandom(MatrixDescriber mx){
 	register int i, j;
     for(i = 0; i < mx.m; i++){
@@ -111,7 +111,7 @@ void fillMatrixWithRandom(MatrixDescriber mx){
     }
 }
 
-//Função que grava numa matriz um valor dando suas coordenadas
+//Funï¿½ï¿½o que grava numa matriz um valor dando suas coordenadas
 bool setElement(MatrixDescriber mx, Coords pos, float value){
     if(!coordIsInsideMatrix(mx, pos)){
         return false;
@@ -121,7 +121,7 @@ bool setElement(MatrixDescriber mx, Coords pos, float value){
     return true;
 }
 
-//Função que remove uma matriz da memória
+//Funï¿½ï¿½o que remove uma matriz da memï¿½ria
 void deleteMatrix(MatrixDescriber *mx){
     register int i;
     for(i = 0; i < mx->m; i++){
@@ -134,7 +134,7 @@ void deleteMatrix(MatrixDescriber *mx){
     
 }
 
-//Função que imprime os valores guardados em uma matriz
+//Funï¿½ï¿½o que imprime os valores guardados em uma matriz
 void printMatrix(MatrixDescriber mx){
     register int i, j;
     for(i = 0; i < mx.m; i++){
